replace hexAlpha with hex
