+++ .github/workflows/consumer_test.yml
@@ -21,7 +21,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - run: |
-          echo "Re-enable this workflow when primer/brand has been updated to usev8 tokens"
+          echo "Re-enable this workflow when primer/brand has been updated to latest tokens"
       # - name: Checkout current repo
       #   uses: actions/checkout@v4
