+++ .github/workflows/visual_regression_test.yml
@@ -14,10 +14,10 @@ jobs:
       fail-fast: false
       matrix:
         shard: [1, 2, 3, 4, 5]
-    
+
     steps:
       - uses: actions/checkout@v4
-  
+
       - name: Setup node
         uses: actions/setup-node@v4
         with:
@@ -27,7 +27,7 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Build v8 tokens
+      - name: Build tokens
         run: npm run build
 
       - name: Build storybook
@@ -68,7 +68,7 @@ jobs:
     if: ${{ always() }}
     runs-on: ubuntu-latest
     needs: vrt-runner
-    
+
     steps:
       - uses: actions/checkout@v4
