+++ src/platforms/javascript.ts
@@ -8,7 +8,6 @@ export const javascript: PlatformInitializer = (outputFile, prefix, buildPath): 
   transforms: [
     'color/hex',
     'color/hexMix',
-    'color/hexAlpha',
     'dimension/rem',
     'shadow/css',
     'border/css',
