+++ src/platforms/json.ts
@@ -8,7 +8,6 @@ export const json: PlatformInitializer = (outputFile, prefix, buildPath): StyleD
   transforms: [
     'color/hex',
     'color/hexMix',
-    'color/hexAlpha',
     'dimension/rem',
     'shadow/css',
     'border/css',
