+++ src/transformers/colorToHex.ts
@@ -2,6 +2,7 @@ import {toHex} from 'color2k'
 import {isColor} from '../filters'
 import type StyleDictionary from 'style-dictionary'
 import {getTokenValue} from './utilities/getTokenValue'
+import {alpha} from './utilities/alpha'
 /**
  * @description converts color tokens value to `hex6` or `hex8`
  * @type value transformer — [StyleDictionary.ValueTransform](https://github.com/amzn/style-dictionary/blob/main/types/Transform.d.ts)
@@ -12,5 +13,11 @@ export const colorToHex: StyleDictionary.Transform = {
   type: `value`,
   transitive: true,
   matcher: isColor,
-  transformer: (token: StyleDictionary.TransformedToken) => toHex(getTokenValue(token)),
+  transformer: (token: StyleDictionary.TransformedToken) => {
+    const alphaValue = token.alpha ?? token.$extensions?.alpha
+    if (alphaValue === null || alphaValue === undefined) {
+      return toHex(getTokenValue(token))
+    }
+    return toHex(alpha(getTokenValue(token), alphaValue, token))
+  },
 }
